redesign(27b/29p): بازطراحیِ جزئیاتِ وام + بررسیِ ایستای کمک‌تابعِ Long-only
بیلدِ ۵۱۴ شکست: بسته‌ی فایلِ ۱ ‏`rialToToman(loan.installment)` نوشته بود، ولی
rialToToman/tomanToRial امضاشون Long ـه و مبلغ‌های پروژه Double ـن. دو قاعده‌ی
درست، برخوردشون خطا. هر دو فراخوان با toLong/toDouble اصلاح شد.

بررسیِ هفدهم (longhelpers.py) همین کلاس خطا رو می‌گیره. moneytypes.py نمی‌گرفت
چون اسمِ پارامترِ خودِ این توابع (rial/toman) تو الگوی پول‌مانندش نیست - و
عوض‌کردنِ امضای خودِ توابع ریسکِ بی‌مورد داشت (فراخوان‌های Longِ درست هم دارن).
سه مثبتِ کاذب (متغیرِ محلیِ Long، ‏toLongOrNull، و it داخلِ زنجیره‌ی toLong) رفع شد.

فایلِ ۲ (LoanDetailScreen) از کلاد دیزاین:
- باگِ اصلی: عددِ «مانده» از loan.installmentِ کهنه حساب می‌شد، نه از ردیف‌ها -
  با جدولِ زیرش جور در نمی‌آمد. حالا جمعِ ردیف‌های پرداخت‌نشده‌ست.
- واحد به تومان، ارقامِ فارسی، حذفِ کارتِ تکراریِ بالای صفحه، حذفِ سومین کپیِ
  نامِ ماهِ جلالی (حالا persianMonthName مشترک)، دیالوگِ حذف با قالبِ واحد.
- دو ناسازگاری با کدِ واقعی اصلاح شد: پارامترهای ConfirmDialog ‏(consequence/
  actionLabel نه body/confirmLabel) و همون تبدیلِ Long/Double.

+++ native-android/app/src/main/kotlin/ir/sadteam/loancalc/ui/myloans/MyLoansScreen.kt
@@ -189,7 +189,7 @@ private fun List<LoanEntity>.sortedByOption(option: LoanSortOption, viewModel: M
 private fun isLoanSettled(loan: LoanEntity) = loan.n > 0 && loan.paidCount >= loan.n
 
 /** ذخیره ریال است و نمایش تومان (بندِ ۲ی README) - تبدیل فقط همین‌جا، لبه‌ی UI. */
-private fun amountToman(rial: Double): String = fmt(rialToToman(rial)).faDigits()
+private fun amountToman(rial: Double): String = fmt(rialToToman(rial.toLong()).toDouble()).faDigits()
 
 private val jalaliMonthNames = listOf(
     "فروردین", "اردیبهشت", "خرداد", "تیر", "مرداد", "شهریور",
@@ -1254,7 +1254,7 @@ private fun DashboardSummary(
                                 val amount = amountText.toDoubleOrNull() ?: 0.0
                                 if (label.trim().isNotEmpty() && amount > 0) {
                                     // ذخیره ریال است، ورودی تومان.
-                                    onAddIncome(label.trim(), tomanToRial(amount), type)
+                                    onAddIncome(label.trim(), tomanToRial(amount.toLong()).toDouble(), type)
                                     label = ""
                                     amountText = ""
                                     showAddIncome = false
